PlayGame.java
This is the code the simulated the whole game. Since it is a game between a computer and the user most of the coding is done in the main method.
It starts with the user entering the position of his ships and grenades, then we use a random method to randomize computer's ships and grenade. Then the game start with the user entering a position where he wants to put his first rocket. If that position is computer's ship then it sinks if it is computer's grenade user losses a turn. Same thing happens for the computer but it's rockets are generated using the random function.
The game ends when all the ships of one of the player are sunk which then invoked the ifWin() method.